use usize

+++ src/engine.rs
@@ -23,7 +23,7 @@ enum Input {
         callback: Box<dyn FnOnce(Option<Lan>) + Send>,
         stop_signal: Arc<AtomicBool>,
     },
-    SetHashSize(u64),
+    SetHashSize(usize),
     ClearHash,
 }
 pub struct Engine {
@@ -115,7 +115,7 @@ impl Engine {
             stop_signal.store(true, Ordering::Relaxed);
         }
     }
-    pub fn set_hash_size(&self, size: u64) {
+    pub fn set_hash_size(&self, size: usize) {
         self.input.send(Input::SetHashSize(size)).unwrap();
     }
     pub fn clear_hash(&self) {
